+++ src/app/theme/index.ts
@@ -1,5 +1,5 @@
 export { BaMenuService } from './services';
-export { InstanceService } from './services';
+export { InstanceService } from '../services';
 export { JobService } from './services';
 export * from './theme.constants';
 export * from './theme.configProvider';
